refactor: validation 위치 변경 및 js 확장자 추가
