+++ src/util/validation.js
@@ -1,0 +1,3 @@
+export const ERROR_COUNT_LABEL = "횟수는 양의 정수만 가능합니다.";
+
+export const ERROR_NAME_LABEL = "이름은 5자 이하만 가능합니다.";
